+++ CHANGELOG.md
@@ -1,3 +1,19 @@
+## [1.11.0](https://github.com/Basis-Theory/basistheory-android/compare/1.10.0...1.11.0) (2023-05-03)
+
+
+### Features
+
+* Adds typeface property to elements ([#59](https://github.com/Basis-Theory/basistheory-android/issues/59)) ([c00241c](https://github.com/Basis-Theory/basistheory-android/commit/c00241ca39d716d8fd0e327d8eb664d096f1d49b))
+
+
+## [1.10.0](https://github.com/Basis-Theory/basistheory-android/compare/1.9.0...1.10.0) (2023-05-02)
+
+
+### Features
+
+* Adds support for hidden input types ([#58](https://github.com/Basis-Theory/basistheory-android/issues/58)) ([98b36f9](https://github.com/Basis-Theory/basistheory-android/commit/98b36f9b928f5826b51de7f0290db4ca10189473))
+
+
 ## [1.9.0](https://github.com/Basis-Theory/basistheory-android/compare/1.8.0...1.9.0) (2023-03-31)